--- atoms/auth.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const authState = atom({
-  key: 'authState',
-  default: '',
-});
