+++ components/common/footer/footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './style.module.css';
+import styles from './footer.module.css';
 
 export default function Footer() {
   const date = new Date();
