--- components/post_request/input/style.module.css
@@ -1,4 +1,0 @@
-.container {
-  width: 100%;
-  margin-bottom: 1rem;
-}
