--- data/home.ts
@@ -1,4 +1,0 @@
-export interface Card {
-  bg: string;
-  msg: Function;
-}
