--- data/request.ts
@@ -1,51 +1,0 @@
-import { District, Suburb } from './user';
-export const jobCategoryList = [
-  'car',
-  'computer',
-  'boat',
-  'housekeeping',
-  'language',
-  'repair',
-  'etc',
-] as const;
-export type JobCategory = typeof jobCategoryList[number];
-
-export type RequestValues = {
-  title: string;
-  district: District;
-  suburb: Suburb;
-  category: JobCategory;
-  detail: string;
-  images: string[];
-};
-export type RequestErrorValues = {
-  title?: string;
-  detail?: string;
-};
-
-export type Image = {
-  file: File;
-  name: string;
-  size: string;
-};
-
-export type RequestsItem = {
-  id: number;
-  title: string;
-  district: string;
-  suburb: string;
-  category: string;
-  detail: string;
-  images: string[];
-  createdAt: string;
-  userId: number;
-  username: string;
-  pay: number;
-};
-
-export type SearchValues = {
-  keyword: string;
-  district: District | 'All Location';
-  suburb: Suburb | '';
-  category: JobCategory | 'All Category';
-};
