+++ package.json
@@ -13,7 +13,6 @@
     "next": "12.0.8",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "recoil": "^0.5.2",
     "swr": "^1.2.2"
   },
   "devDependencies": {
