+++ pages/sign_in.tsx
@@ -1,6 +1,5 @@
 import { useRouter } from 'next/router';
 import React, { useContext, useEffect, useState } from 'react';
-import AuthService from '../utils/modules/auth';
 import styles from '../styles/sign_in.module.css';
 import { authContext } from '../context/auth';
